Move security transport internals into services
Move authentication, authorization, external credential, and secrets provider transports into their owning service packages while leaving internal/providerhost as compatibility wrappers.

New service interfaces:

authenticationProvider, err := authentication.NewExecutable(ctx, authentication.ExecConfig{Name: "auth", CallbackURL: callbackURL, SessionKey: sessionKey})

authorizationProvider, err := authorization.NewExecutable(ctx, authorization.ExecConfig{Name: "authz", HostServices: hostServices})

proto.RegisterAuthorizationProviderServer(srv, authorization.NewProviderServer(provider))

externalCredentials, err := externalcredentials.NewExecutable(ctx, externalcredentials.ExecConfig{Name: "credentials", HostServices: hostServices})

proto.RegisterExternalCredentialProviderServer(srv, externalcredentials.NewProviderServer(provider))

secretManager, err := secrets.NewExecutable(ctx, secrets.ExecConfig{Name: "secrets"})

Compatibility wrappers still work through providerhost, for example providerhost.NewExecutableAuthorizationProvider(ctx, providerhost.AuthorizationExecConfig{Name: "authz"}).

Verification:

go test ./...

golangci-lint run ./services/authentication ./services/authorization ./services/externalcredentials ./services/secrets ./internal/providerhost

git diff --cached --check
