+++ gestaltd/internal/providerhost/authorization_server.go
@@ -1,46 +1,17 @@
 package providerhost
 
 import (
-	"context"
-
 	proto "github.com/valon-technologies/gestalt/sdk/go/gen/v1"
 	"github.com/valon-technologies/gestalt/server/core"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/emptypb"
+	authorizationservice "github.com/valon-technologies/gestalt/server/services/authorization"
 )
 
-const DefaultAuthorizationSocketEnv = "GESTALT_AUTHORIZATION_SOCKET"
+const DefaultAuthorizationSocketEnv = authorizationservice.DefaultSocketEnv
 
 func AuthorizationSocketTokenEnv() string {
-	return DefaultAuthorizationSocketEnv + "_TOKEN"
-}
-
-type authorizationProviderServer struct {
-	proto.UnimplementedAuthorizationProviderServer
-	provider core.AuthorizationProvider
+	return authorizationservice.SocketTokenEnv()
 }
 
 func NewAuthorizationProviderServer(provider core.AuthorizationProvider) proto.AuthorizationProviderServer {
-	return &authorizationProviderServer{provider: provider}
-}
-
-func (s *authorizationProviderServer) SearchSubjects(ctx context.Context, req *proto.SubjectSearchRequest) (*proto.SubjectSearchResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "request is required")
-	}
-	resp, err := s.provider.SearchSubjects(ctx, req)
-	if err != nil {
-		return nil, status.Errorf(codes.Unknown, "search subjects: %v", err)
-	}
-	if resp == nil {
-		return nil, status.Error(codes.Internal, "authorization provider returned nil response")
-	}
-	return resp, nil
-}
-
-func (s *authorizationProviderServer) GetMetadata(context.Context, *emptypb.Empty) (*proto.AuthorizationMetadata, error) {
-	return &proto.AuthorizationMetadata{
-		Capabilities: []string{"search_subjects"},
-	}, nil
+	return authorizationservice.NewProviderServer(provider)
 }
